+++ src/models/dto/index.ts
@@ -1,7 +1,5 @@
 export * from './paginated-response.dto';
-export * from './notification-auth.dto';
 export * from './user.dto';
 export * from './paginated-request.dto';
-export * from './notification-auth.dto';
 export * from './message-author.dto';
 export * from './message-notification.dto';
