+++ src/sockets/notification-auth.dto.ts
@@ -1,15 +1,13 @@
 import { IsEnum, IsString } from 'class-validator';
-import { ApiProperty } from '@nestjs/swagger';
-import { AuthProviders } from '../';
+import { AuthProviders } from 'src/models';
+
 
 /** DTO for authenticating with the notification gateway */
 export class NotificationAuthDTO {
   /** Provider to authenticate with */
-  @ApiProperty()
   @IsEnum(AuthProviders)
   provider: AuthProviders;
   /** Uid of user */
   @IsString()
-  @ApiProperty()
   id: string;
 }
